Copy useBackgroundQuery defer tests to own file
