+++ src/react/hooks/__tests__/useBackgroundQuery.test.tsx
@@ -30,7 +30,6 @@ import {
   NetworkStatus,
 } from "@apollo/client";
 import { InMemoryCache } from "@apollo/client/cache";
-import { Defer20220824Handler } from "@apollo/client/incremental";
 import type { QueryRef } from "@apollo/client/react";
 import {
   ApolloProvider,
@@ -1391,150 +1390,6 @@ it("works with startTransition to change variables", async () => {
   }
 });
 
-it('does not suspend deferred queries with data in the cache and using a "cache-and-network" fetch policy', async () => {
-  interface Data {
-    greeting: {
-      __typename: string;
-      message: string;
-      recipient: { name: string; __typename: string };
-    };
-  }
-
-  const query: TypedDocumentNode<Data> = gql`
-    query {
-      greeting {
-        message
-        ... on Greeting @defer {
-          recipient {
-            name
-          }
-        }
-      }
-    }
-  `;
-
-  const link = new MockSubscriptionLink();
-  const cache = new InMemoryCache();
-  cache.writeQuery({
-    query,
-    data: {
-      greeting: {
-        __typename: "Greeting",
-        message: "Hello cached",
-        recipient: { __typename: "Person", name: "Cached Alice" },
-      },
-    },
-  });
-  const client = new ApolloClient({
-    cache,
-    link,
-    incrementalHandler: new Defer20220824Handler(),
-  });
-
-  const renderStream = createDefaultProfiler<Data>();
-
-  const { SuspenseFallback, ReadQueryHook } =
-    createDefaultTrackedComponents(renderStream);
-
-  function App() {
-    useTrackRenders();
-    const [queryRef] = useBackgroundQuery(query, {
-      fetchPolicy: "cache-and-network",
-    });
-
-    return (
-      <Suspense fallback={<SuspenseFallback />}>
-        <ReadQueryHook queryRef={queryRef} />
-      </Suspense>
-    );
-  }
-
-  using _disabledAct = disableActEnvironment();
-  await renderStream.render(<App />, { wrapper: createClientWrapper(client) });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([App, ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello cached",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.loading,
-    });
-  }
-
-  link.simulateResult({
-    result: {
-      data: {
-        greeting: { __typename: "Greeting", message: "Hello world" },
-      },
-      hasNext: true,
-    },
-  });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "streaming",
-      error: undefined,
-      networkStatus: NetworkStatus.streaming,
-    });
-  }
-
-  link.simulateResult(
-    {
-      result: {
-        incremental: [
-          {
-            data: {
-              recipient: { name: "Alice", __typename: "Person" },
-              __typename: "Greeting",
-            },
-            path: ["greeting"],
-          },
-        ],
-        hasNext: false,
-      },
-    },
-    true
-  );
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.ready,
-    });
-  }
-
-  await expect(renderStream).not.toRerender({ timeout: 50 });
-});
 
 it("reacts to cache updates", async () => {
   const { query, mocks } = setupSimpleCase();
@@ -3816,158 +3671,6 @@ it('suspends and does not use partial data when changing variables and using a "
   await expect(renderStream).not.toRerender({ timeout: 50 });
 });
 
-it('does not suspend deferred queries with partial data in the cache and using a "cache-first" fetch policy with `returnPartialData`', async () => {
-  interface QueryData {
-    greeting: {
-      __typename: string;
-      message?: string;
-      recipient?: {
-        __typename: string;
-        name: string;
-      };
-    };
-  }
-
-  const query: TypedDocumentNode<QueryData> = gql`
-    query {
-      greeting {
-        message
-        ... on Greeting @defer {
-          recipient {
-            name
-          }
-        }
-      }
-    }
-  `;
-
-  const link = new MockSubscriptionLink();
-  const cache = new InMemoryCache();
-
-  // We are intentionally writing partial data to the cache. Supress console
-  // warnings to avoid unnecessary noise in the test.
-  {
-    using _consoleSpy = spyOnConsole("error");
-    cache.writeQuery({
-      query,
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-    });
-  }
-
-  const client = new ApolloClient({
-    link,
-    cache,
-    incrementalHandler: new Defer20220824Handler(),
-  });
-
-  const renderStream = createDefaultProfiler<DeepPartial<QueryData>>();
-  const { SuspenseFallback, ReadQueryHook } =
-    createDefaultTrackedComponents(renderStream);
-
-  function App() {
-    useTrackRenders();
-    const [queryRef] = useBackgroundQuery(query, {
-      fetchPolicy: "cache-first",
-      returnPartialData: true,
-    });
-
-    return (
-      <Suspense fallback={<SuspenseFallback />}>
-        <ReadQueryHook queryRef={queryRef} />
-      </Suspense>
-    );
-  }
-
-  using _disabledAct = disableActEnvironment();
-  await renderStream.render(<App />, { wrapper: createClientWrapper(client) });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([App, ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "partial",
-      error: undefined,
-      networkStatus: NetworkStatus.loading,
-    });
-  }
-
-  link.simulateResult({
-    result: {
-      data: {
-        greeting: { message: "Hello world", __typename: "Greeting" },
-      },
-      hasNext: true,
-    },
-  });
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Cached Alice" },
-        },
-      },
-      dataState: "streaming",
-      error: undefined,
-      networkStatus: NetworkStatus.streaming,
-    });
-  }
-
-  link.simulateResult(
-    {
-      result: {
-        incremental: [
-          {
-            data: {
-              __typename: "Greeting",
-              recipient: { name: "Alice", __typename: "Person" },
-            },
-            path: ["greeting"],
-          },
-        ],
-        hasNext: false,
-      },
-    },
-    true
-  );
-
-  {
-    const { snapshot, renderedComponents } = await renderStream.takeRender();
-
-    expect(renderedComponents).toStrictEqual([ReadQueryHook]);
-    expect(snapshot.result).toStrictEqualTyped({
-      data: {
-        greeting: {
-          __typename: "Greeting",
-          message: "Hello world",
-          recipient: { __typename: "Person", name: "Alice" },
-        },
-      },
-      dataState: "complete",
-      error: undefined,
-      networkStatus: NetworkStatus.ready,
-    });
-  }
-
-  await expect(renderStream).not.toRerender({ timeout: 50 });
-});
 
 it.each<useBackgroundQuery.FetchPolicy>([
   "cache-first",
